refactor: change dao to repository
