+++ src/main/java/kitchenpos/dao/TableGroupRepository.java
@@ -4,8 +4,9 @@ import kitchenpos.domain.TableGroup;
 
 import java.util.List;
 import java.util.Optional;
+import org.springframework.data.repository.Repository;
 
-public interface TableGroupDao {
+public interface TableGroupRepository extends Repository<TableGroup, Long> {
     TableGroup save(TableGroup entity);
 
     Optional<TableGroup> findById(Long id);
